test/requirements.txt
args==0.1.0
asyncio-nats-client==0.11.5
common==0.1.2
nats-py==2.3.1
python-dotenv==1.0.0
pluggy==1.3.0
pytest==7.4.0